requirements.txt
absl-py==0.13.0
Adafruit-Blinka==6.13.0
adafruit-circuitpython-busdevice==5.0.6
Adafruit-PlatformDetect==3.15.3
Adafruit-PureIO==1.1.9
astunparse==1.6.3
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
colorzero==2.0
flatbuffers==1.12
gast==0.4.0
google-auth==1.34.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
gpiozero==1.6.2
grpcio==1.34.1
h5py==3.1.0
idna==3.2
imutils==0.5.4
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
Markdown==3.3.4
numpy==1.19.5
oauthlib==3.1.1
opencv-contrib-python==4.5.3.56
opt-einsum==3.3.0
pkg_resources==0.0.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyftdi==0.53.2
PyMLX90614==0.0.3
pyserial==3.5
pyusb==1.1.0
requests==2.26.0
requests-oauthlib==1.3.0
RPi.GPIO==0.7.0
rsa==4.7.2
six==1.15.0
smbus2==0.4.1
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
